Apply hover effect over divs and adjust grid size
Added event listener for 'mouseover' on the container div and set it
so that the background color for target element changes to blue.

Initially, the size of the container div depended upon the size of the
unit divs when it should have been the other way round. Hence, the size
property of unit divs has been removed and has been applied to the container
div. Now the unit divs auto-adjust their size in relation to the container
div because of flexbox.