Add button to set grid-size and add functions to create & delete grid
Initially the grid had a fixed size of 16. Hence added a size button so
that the user can set size of their choice (limited to size of 100 to
avoid delay, freeze, etc.)

Created two size variables; one to store current size and another to store
new size.

Created a createGrid() function that takes size as input and creates a
new grid of desired size. No changes made to algorithm.

Created a deleteGrid() function that takes size as input and removes the
existing grid (by deleting the rows) to make space for new grid.