Implement progressive darkening effect to hovering
Assigned a value of 1 to each unit div to serve as brightness value as
it is tedious to access the
brightness attribute -> (element.style.filter = 'brightness()') manually.

Created a darken() function which when called, lowers the brightness
of unit div by 10% upon each interaction, hence leading to completely
black unit div in 10 interactions.

Hence the value attribute is decreased by 0.1 upon each interaction
along with the brightness attribute.